Update form for flight
